app/requirements.txt
# MIT License

click~=8.1
pyyaml~=6.0
aioprocessing~=2.0
numpy~=1.24
scipy~=1.10
